// packages/app/src/lib/react-native-radial-slider/src/components/RadialSlider/types.ts
export type RadialSliderProps = {
  /**
   * Radious of radial slider.
   */
  radius?: number;
  /**
   * Min value of radial slider.
   */
  min?: number;
  /**
   * Max value to radial slider.
   */
  max?: number;
  /**
   * Radial slider step value.
   */
  step?: number;
  /**
   * Show selection upto this value.
   */
  value?: number;
  /**
   * Radial slider unit.
   */
  unit?: string;
  /**
   * Radious of thumb/knob.
   */
  thumbRadius?: number;
  /**
   * Color of thumb/knob.
   */
  thumbColor?: string;
  /**
   * Border width of thumb/knob.
   */
  thumbBorderWidth?: number;
  /**
   * Border color of thumb/knob.
   */
  thumbBorderColor?: string;
  /**
   * Size of marker line.
   */
  markerLineSize?: number;
  /**
   * Width of slider.
   */
  sliderWidth?: number;
  /**
   * Color of unselected slider track.
   */
  sliderTrackColor?: string;
  /**
   * Color of unselected lines.
   */
  lineColor?: string;
  /**
   * Space between each line.
   */
  lineSpace?: number;
  /**
   * Gradient color of selected track.
   */
  linearGradient?: {offset: string; color: string}[];
  linearGradientPlaceholder?: {offset: string; color: string}[];
  /**
   * Callback function which fired on change in track.
   */
  onChange?: (v: number) => void;
  /**
   * Callback function which defines what to do after completion.
   */
  onComplete?: (v: number) => void;
  /**
   * Center content styling.
   */
  centerContentStyle?: StyleProp<ViewStyle>;
  /**
   * Status title container styling.
   */
  titleStyle?: StyleProp<TextStyle>;
  /**
   * Status subtitle text styling.
   */
  subTitleStyle?: StyleProp<TextStyle>;
  /**
   * Center value style.
   */
  valueStyle?: StyleProp<TextStyle>;
  /**
   * Button container styling.
   */
  buttonContainerStyle?: StyleProp<ViewStyle>;
  /**
   * Left Icon styling.
   */
  leftIconStyle?: StyleProp<ViewStyle>;
  /**
   * Right Icon styling.
   */
  rightIconStyle?: StyleProp<ViewStyle>;
  /**
   * Whole content styling.
   */
  contentStyle?: StyleProp<ViewStyle>;
  /**
   * Unit text styling.
   */
  unitStyle?: StyleProp<TextStyle>;
  /**
   * Inner container styling.
   */
  style?: StyleProp<ViewStyle>;
  /**
   * Radian of the slider.
   */
  openingRadian?: number;
  /**
   * If true, buttons will be in disabled state.
   */
  disabled?: boolean;
  /**
   * If true, slider will be hidden.
   */
  isHideSlider?: boolean;
  /**
   * If true, circle will be hidden.
   */
  isHideCircle?: boolean;
  /**
   * If true, center content will be hidden.
   */
  isHideCenterContent?: boolean;
  /**
   * If true, title will be hidden.
   */
  isHideTitle?: boolean;
  /**
   * If true, subtitle will be hidden.
   */
  isHideSubtitle?: boolean;
  /**
   * If true, value will be hidden.
   */
  isHideValue?: boolean;
  /**
   * If true, tail text will be hidden.
   */
  isHideTailText?: boolean;
  /**
   * If true, buttons will be hidden.
   */
  isHideButtons?: boolean;
  /**
   * If true, lines will be hidden.
   */
  isHideLines?: boolean;
  /**
   * If true, marked lines will be hidden.
   */
  isHideMarkerLine?: boolean;
  /**
   * If true, marked value will be hidden.
   */
  fixedMarker?: boolean;
  /**
   * Color for icon
   */
  stroke?: string;

  children?: ReactNode;
};